+++ frontend/src/app/mocks/services/policy-mock/policy.model.ts
@@ -1,3 +1,5 @@
+import { OperatorType, Policy, PolicyAction, PolicyResponseMap } from '@page/policies/model/policy.model';
+
 /********************************************************************************
  * Copyright (c) 2022, 2023, 2024 Contributors to the Eclipse Foundation
  *
@@ -17,36 +19,183 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 // For now Mocks are built by current response (any). This is because Policy Model changes frequently
-export const getPolicies =  () => [
+export const getPolicies = (): Policy[] => {
+  return mockedPolicyList;
+};
+
+export const getPolicyById = (policyId: string | ReadonlyArray<string>): Policy => {
+  return mockedPolicyList.filter(policy => policy.policyId === policyId)[0];
+};
+
+
+export const mockedPolicyList: Policy[] = [
   {
-    policyId: "Mocked_Policy_Id",
-    createdOn: 1706690077.834424001,
-    validUntil: 1727740799.990000000,
-    permissions: [
+    'policyId': 'default-policy',
+    'bpn': 'BPNL00000003CML1,BPNL00000003CNKC',
+    'createdOn': '2024-05-29T06:18:40Z',
+    'validUntil': '2029-05-29T06:18:40Z',
+    'permissions': [
       {
-        action: "USE",
-        constraints: [
-          {
+        'action': PolicyAction.USE,
+        'constraints': {
+          'and': null,
+          'or': [
+            {
+              'leftOperand': 'cx-policy:FrameworkAgreement',
+              'operatorTypeResponse': OperatorType.EQ,
+              'rightOperand': 'traceability:1.0',
+            },
+            {
+              'leftOperand': 'cx-policy:UsagePurpose',
+              'operatorTypeResponse': OperatorType.EQ,
+              'rightOperand': 'cx.core.industrycore:1',
+            },
+          ],
+        },
+      },
+    ],
+  },
+  {
+    'policyId': 'default-policy-2',
+    'bpn': 'BPNL00000003CML1',
+    'createdOn': '2024-05-29T06:18:40Z',
+    'validUntil': '2029-05-29T06:18:40Z',
+    'permissions': [
+      {
+        'action': PolicyAction.USE,
+        'constraints': {
+          'and': [
+            {
+              'leftOperand': 'cx-policy:FrameworkAgreement',
+              'operatorTypeResponse': OperatorType.EQ,
+              'rightOperand': 'traceability:1.0',
+            },
+            {
+              'leftOperand': 'cx-policy:UsagePurpose',
+              'operatorTypeResponse': OperatorType.EQ,
+              'rightOperand': 'cx.core.industrycore:1',
+            },
+          ],
+          'or': null,
+        },
+      },
+    ],
+  },
+];
+
+const mockedPolicies = {
+  page: 0,
+  pageCount: 0,
+  pageSize: 10,
+  totalItems: 2,
+
+  content: [
+    {
+      bpnSelection: [ 'BPN10000000OEM0A', 'BPN10000000OEM0B' ],
+      policyName: 'Mocked_Policy_Name_1',
+      policyId: 'Mocked_Policy_1',
+      accessType: PolicyAction.ACCESS,
+      createdOn: '2024-05-29T06:18:40Z',
+      validUntil: '2024-05-29T06:18:40Z',
+      constraints: [ 'Membership = active', 'AND FrameworkAgreement.traceability in [active]', 'AND PURPOSE = ID 3.1 Trace' ],
+      permissions: [
+        {
+          action: PolicyAction.USE,
+          constraint: {
             and: [
               {
-                leftOperand: "PURPOSE",
-                operatorTypeResponse: "EQ",
-                rightOperands: [
-                  "ID 3.0 Trace"
-                ]
+                leftOperand: 'PURPOSE',
+                operator: {
+                  id: OperatorType.EQ,
+                },
+                rightOperand: 'ID 3.0 Trace',
               },
-              {
-                leftOperand: "PURPOSE",
-                operatorTypeResponse: "EQ",
-                rightOperands: [
-                  "ID 3.0 Trace"
-                ]
-              }
             ],
-            or: []
-          }
-        ]
-      }
-    ]
-  }
-]
+            or: [
+              {
+                leftOperand: 'PURPOSE',
+                operator: {
+                  id: OperatorType.EQ,
+                },
+                rightOperand: 'ID 3.0 Trace',
+              },
+            ],
+          },
+        },
+      ],
+    },
+    {
+      bpnSelection: [ 'BPN10000000OEM0A', 'BPN10000000OEM0B' ],
+      policyName: 'Mocked_Policy_Name_2',
+      policyId: 'Mocked_Policy_2',
+      accessType: PolicyAction.USE,
+      createdOn: '2024-05-29T06:18:40Z',
+      validUntil: '2024-05-29T06:18:40Z',
+      constraints: [ 'PURPOSE = ID 3.1 Trace', 'OR PURPOSE = ID 3.0 Trace' ],
+      permissions: [
+        {
+          action: PolicyAction.USE,
+          constraint: {
+            and: [
+              {
+                leftOperand: 'PURPOSE',
+                operator: {
+                  id: OperatorType.EQ,
+                },
+                rightOperand: 'BMW',
+              },
+            ],
+            or: [
+              {
+                leftOperand: 'PURPOSE',
+                operator: {
+                  id: OperatorType.EQ,
+                },
+                rightOperand: 'ID 3.0 Trace',
+              },
+            ],
+          },
+        },
+      ],
+    },
+
+  ],
+
+};
+
+export const MockPolicyResponseMap: PolicyResponseMap = {
+  'default': [
+    {
+      'validUntil': '2024-06-30T11:07:00Z',
+      'payload': {
+        '@context': {
+          'odrl': 'http://www.w3.org/ns/odrl/2/',
+        },
+        '@id': 'asdadasdas',
+        'policy': {
+          'policyId': 'asdadasdas',
+          'createdOn': '2024-06-13T09:07:32.229901783Z',
+          'validUntil': '2024-06-30T11:07:00Z',
+          'permissions': [
+            {
+              'action': PolicyAction.USE,
+              'constraint': {
+                'and': null,
+                'or': [
+                  {
+                    'leftOperand': 'asd',
+                    'operator': {
+                      '@id': OperatorType.EQ,
+                    },
+                    'odrl:rightOperand': 'dsa',
+                  },
+                ],
+              },
+            },
+          ],
+        },
+      },
+    },
+  ],
+};
+