+++ frontend/src/app/modules/page/admin/presentation/policy-management/policy-editor/policy-data.ts
@@ -1,0 +1,36 @@
+import { ConstraintLogicType, OperatorType } from '@page/policies/model/policy.model';
+
+export const OperatorTypesAsSelectOptionsList: any[] = Object.keys(OperatorType).map(key => {
+  let convertedOperatorType = '';
+  switch (key) {
+    case OperatorType.EQ:
+      convertedOperatorType = '=';
+      break;
+    case OperatorType.NEQ:
+      convertedOperatorType = '!=';
+      break;
+    case OperatorType.LT:
+      convertedOperatorType = '<';
+      break;
+    case OperatorType.GT:
+      convertedOperatorType = '>';
+      break;
+    case OperatorType.LTEQ:
+      convertedOperatorType = '<=';
+      break;
+    case OperatorType.GTEQ:
+      convertedOperatorType = '>=';
+      break;
+    default:
+      convertedOperatorType = key;
+  }
+  return {
+    label: convertedOperatorType, value: convertedOperatorType,
+  };
+});
+
+export const ConstraintLogicTypeAsSelectOptionsList: any[] = Object.keys(ConstraintLogicType).map(key => {
+  return {
+    label: key, value: key,
+  };
+});