+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/notification/domain/base/service/NotificationsEDCFacade.java
@@ -45,6 +45,7 @@ import org.eclipse.tractusx.traceability.notification.domain.base.model.Notifica
 import org.eclipse.tractusx.traceability.notification.domain.base.model.NotificationMessage;
 import org.eclipse.tractusx.traceability.notification.domain.base.model.NotificationStatus;
 import org.eclipse.tractusx.traceability.notification.domain.base.model.NotificationType;
+import org.eclipse.tractusx.traceability.notification.domain.notification.exception.NotificationNotFoundException;
 import org.eclipse.tractusx.traceability.notification.domain.notification.repository.NotificationRepository;
 import org.eclipse.tractusx.traceability.notification.infrastructure.edc.model.EDCNotification;
 import org.eclipse.tractusx.traceability.notification.infrastructure.edc.model.EDCNotificationFactory;
@@ -90,24 +91,24 @@ public class NotificationsEDCFacade {
     private static final String CX_TAXO_QUALITY_ALERT_UPDATE = "https://w3id.org/catenax/taxonomy#UpdateQualityAlertNotification";
 
     public void startEdcTransfer(
-            final NotificationMessage notification,
+            final NotificationMessage notificationMessage,
             final String receiverEdcUrl,
             final String senderEdcUrl) {
 
-        CatalogItem catalogItem = getCatalogItem(notification, receiverEdcUrl);
+        CatalogItem catalogItem = getCatalogItem(notificationMessage, receiverEdcUrl);
 
-        String contractAgreementId = negotiateContractAgreement(receiverEdcUrl, catalogItem, notification.getSendTo());
+        String contractAgreementId = negotiateContractAgreement(receiverEdcUrl, catalogItem, notificationMessage.getSentTo());
 
         final EndpointDataReference dataReference = endpointDataReferenceStorage.get(contractAgreementId)
                 .orElseThrow(() -> new NoEndpointDataReferenceException("No EndpointDataReference was found"));
 
-        notification.setContractAgreementId(contractAgreementId);
+        notificationMessage.setContractAgreementId(contractAgreementId);
 
         try {
-            EdcNotificationRequest notificationRequest = toEdcNotificationRequest(notification, senderEdcUrl, dataReference);
-            sendRequest(notificationRequest, notification);
+            EdcNotificationRequest notificationRequest = toEdcNotificationRequest(notificationMessage, senderEdcUrl, dataReference);
+            sendRequest(notificationRequest, notificationMessage);
         } catch (Exception e) {
-            throw new SendNotificationException("Failed to send notification.", e);
+            throw new SendNotificationException("Failed to send notificationMessage.", e);
         }
     }
 
@@ -141,7 +142,7 @@ public class NotificationsEDCFacade {
                             CatalogRequest.Builder.newInstance()
                                     .protocol(DEFAULT_PROTOCOL)
                                     .counterPartyAddress(receiverEdcUrl + edcProperties.getIdsPath())
-                                    .counterPartyId(notification.getSendTo())
+                                    .counterPartyId(notification.getSentTo())
                                     .querySpec(QuerySpec.Builder.newInstance()
                                             // https://github.com/eclipse-tractusx/traceability-foss/issues/978
                                             // Probably:
@@ -157,7 +158,7 @@ public class NotificationsEDCFacade {
                     .filter(catalogItem -> {
                         log.info("-- catalog item check --");
                         log.info("Item {}: {}", catalogItem.getItemId(), catalogItem);
-                        boolean isValid = policyCheckerService.isValid(catalogItem.getPolicy(), notification.getSendTo()
+                        boolean isValid = policyCheckerService.isValid(catalogItem.getPolicy(), notification.getSentTo()
                         );
                         log.info("IsValid : {}", isValid);
                         return isValid;
@@ -172,22 +173,30 @@ public class NotificationsEDCFacade {
 
     // TODO this method should be completly handled by EDCNotificationFactory.createEdcNotification which is part of this method currently
     private EdcNotificationRequest toEdcNotificationRequest(
-            final NotificationMessage notification,
+            final NotificationMessage notificationMessage,
             final String senderEdcUrl,
             final EndpointDataReference dataReference
     ) throws JsonProcessingException {
-        EDCNotification edcNotification = EDCNotificationFactory.createEdcNotification(senderEdcUrl, notification);
-        objectMapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-        String body = objectMapper.writeValueAsString(edcNotification);
 
-        HttpHeaders headers = new HttpHeaders();
-        headers.set(Objects.requireNonNull(dataReference.getAuthKey()), dataReference.getAuthCode());
-        headers.set("Content-Type", "application/json");
-        log.info(":::: Send notification Data  body :{}, dataReferenceEndpoint :{}", body, dataReference.getEndpoint());
-        return EdcNotificationRequest.builder()
-                .url(dataReference.getEndpoint())
-                .body(body)
-                .headers(headers).build();
+        Optional<Notification> optionalNotificationById = notificationRepository.findByEdcNotificationId(notificationMessage.getEdcNotificationId());
+
+        if (optionalNotificationById.isPresent()) {
+            EDCNotification edcNotification = EDCNotificationFactory.createEdcNotification(senderEdcUrl, notificationMessage, optionalNotificationById.get());
+            objectMapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
+            String body = objectMapper.writeValueAsString(edcNotification);
+
+            HttpHeaders headers = new HttpHeaders();
+            headers.set(Objects.requireNonNull(dataReference.getAuthKey()), dataReference.getAuthCode());
+            headers.set("Content-Type", "application/json");
+            log.info(":::: Send notificationMessage Data  body :{}, dataReferenceEndpoint :{}", body, dataReference.getEndpoint());
+            return EdcNotificationRequest.builder()
+                    .url(dataReference.getEndpoint())
+                    .body(body)
+                    .headers(headers).build();
+        } else {
+            throw new NotificationNotFoundException("Could not find notification.");
+        }
+
     }